requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
psycopg==3.2.4
pypdf==5.3.0
pytesseract==0.3.13
sqlparse==0.5.3
typing_extensions==4.12.2
